manage inventory
