+++ src/components/dashboard/inventory/purchase/purchase-pdf/purchase-view-dialog.tsx
@@ -17,7 +17,7 @@ import PurchasePdfDoc from "./purchase-pdf-doc";
 import { Close } from "@mui/icons-material";
 import { CompanyDto } from "@/types/company.types";
 import { PurchaseOrder } from "@/types/purchase.types";
-import {PurchaseViewDialogProps } from "@/modules/inventory/types";
+import {PurchaseViewDialogProps } from "@/modules/inventory/types/purchase.types";
 
 const Transition = forwardRef(function Transition(
   props: TransitionProps & {
