+++ src/modules/inventory/types/purchase.types.ts
